+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added isEmpty & isNotEmpty properties to Event2RegistrationDetails [#4043](https://github.com/rokwire/illinois-app/issues/4043).
 - Added events2 query deep link [#4041](https://github.com/rokwire/illinois-app/issues/4041).
 - Exposed scope in content attributes [#4029](https://github.com/rokwire/illinois-app/issues/4029).
+- Implemented conditional content attributes [#4134](https://github.com/rokwire/illinois-app/issues/4134).
 ### Fixed
 - Fixed client paramter in loadEventsEx invocation from loadEvents API.
 - Make sure to always return successfully refreshed token in Auth2 service.